check the remote system for os information. first use md5sum, if that fails, use the perl variant

+++ lib/Rex/Commands/MD5.pm
@@ -63,43 +63,46 @@ sub md5 {
 
     Rex::Logger::debug("Calculating Checksum (md5) of $file");
 
-    my $script = q|
-    use Digest::MD5;
-    print Digest::MD5::md5_hex(<>) . "\n";
-    |;
-
-    my $rnd_file = get_tmp_file;
-
-    my $fh = Rex::Interface::File->create;
-    $fh->open(">", $rnd_file);
-    $fh->write($script);
-    $fh->close;
-
     my $exec = Rex::Interface::Exec->create;
     my $md5;
 
-    if(Rex::is_local() && $^O =~ m/^MSWin/) {
-      $md5 = $exec->exec("perl $rnd_file \"$file\"");
+    my $os = $exec->exec("uname -s");
+    if($os =~ /bsd/i) {
+      $md5 = $exec->exec("/sbin/md5 -q '$file'");
     }
     else {
-      $md5 = $exec->exec("perl $rnd_file '$file'");
+      ($md5) = split(/\s/, $exec->exec("md5sum '$file'"));
     }
 
-    unless($? == 0) {
-      if($^O =~ /bsd/) {
-        $md5 = $exec->exec("/sbin/md5 -q '$file'");
+    if($? != 0) {
+
+      my $script = q|
+      use Digest::MD5;
+      print Digest::MD5::md5_hex(<>) . "\n";
+      |;
+
+      my $rnd_file = get_tmp_file;
+
+      my $fh = Rex::Interface::File->create;
+      $fh->open(">", $rnd_file);
+      $fh->write($script);
+      $fh->close;
+
+      if(Rex::is_local() && $^O =~ m/^MSWin/) {
+        $md5 = $exec->exec("perl $rnd_file \"$file\"");
       }
       else {
-        ($md5) = split(/\s/, $exec->exec("md5sum '$file'"));
+        $md5 = $exec->exec("perl $rnd_file '$file'");
       }
+
+      unless($? == 0) {
+        Rex::Logger::info("Unable to get md5 sum of $file");
+        die("Unable to get md5 sum of $file");
+      }
+
+      Rex::Interface::Fs->create->unlink($rnd_file);
     }
 
-    unless($? == 0) {
-      Rex::Logger::info("Unable to get md5 sum of $file");
-      die("Unable to get md5 sum of $file");
-    }
-
-    Rex::Interface::Fs->create->unlink($rnd_file);
 
     Rex::Logger::debug("MD5SUM ($file): $md5");
     $md5 =~ s/[\r\n]//gms;
